Convert entire library into header-only
Convert the entire library into header-only, so that it can be included
in C++ codebases more easily.
