+++ .travis.yml
@@ -7,5 +7,5 @@ compiler:
 env:
   - CXXFLAGS="-Wall -Werror"
 script:
-  - cmake -DCMAKE_BUILD_TYPE=Debug .
+  - cmake -DCMAKE_BUILD_TYPE=Debug -DPEELO_PROMPT_BUILD_EXAMPLE=ON .
   - cmake --build . -- VERBOSE=1
