+++ CMakeLists.txt
@@ -1,57 +1,82 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 3.0)
-PROJECT(peelocpp_prompt CXX)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.12)
 
-INCLUDE(GNUInstallDirs)
+PROJECT(
+  PeeloPrompt
+  VERSION 0.3.0
+  DESCRIPTION "Header only line editing library based on Linenoise."
+  HOMEPAGE_URL "https://github.com/peelonet/peelo-prompt"
+  LANGUAGES CXX
+)
 
 OPTION(
-  PEELOCPP_PROMPT_BUILD_EXAMPLE
+  PEELO_PROMPT_BUILD_EXAMPLE
   "Whether the example application should be built or not."
   OFF
 )
 
-ADD_LIBRARY(
-  peelocpp_prompt
-  src/prompt.cpp
-)
+INCLUDE(GNUInstallDirs)
 
-TARGET_COMPILE_OPTIONS(
-  peelocpp_prompt
-  PRIVATE
-    -Wall -Werror
+ADD_LIBRARY(${PROJECT_NAME} INTERFACE)
+
+TARGET_INCLUDE_DIRECTORIES(
+  ${PROJECT_NAME}
+  INTERFACE
+    $<BUILD_INTERFACE:${${PROJECT_NAME}_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
 )
 
 TARGET_COMPILE_FEATURES(
-  peelocpp_prompt
-  PRIVATE
+  ${PROJECT_NAME}
+  INTERFACE
     cxx_std_17
-)
-
-TARGET_INCLUDE_DIRECTORIES(
-  peelocpp_prompt
-  PUBLIC
-    $<INSTALL_INTERFACE:include>
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
 )
 
 INSTALL(
   TARGETS
-    peelocpp_prompt
+    ${PROJECT_NAME}
   EXPORT
-    peelocpp_prompt-export
-  LIBRARY DESTINATION
-    ${CMAKE_INSTALL_LIBDIR}
+    ${PROJECT_NAME}_Targets
   ARCHIVE DESTINATION
     ${CMAKE_INSTALL_LIBDIR}
-)
-INSTALL(
-  DIRECTORY
-    ${CMAKE_CURRENT_SOURCE_DIR}/include/peelo
-    DESTINATION
-      include
-    FILES_MATCHING PATTERN
-      "*.hpp"
+  LIBRARY DESTINATION
+    ${CMAKE_INSTALL_LIBDIR}
+  RUNTIME DESTINATION
+    ${CMAKE_INSTALL_BINDIR}
 )
 
-IF(PEELOCPP_PROMPT_BUILD_EXAMPLE)
+INCLUDE(CMakePackageConfigHelpers)
+WRITE_BASIC_PACKAGE_VERSION_FILE(
+  "${PROJECT_NAME}ConfigVersion.cmake"
+  VERSION
+    ${PROJECT_VERSION}
+  COMPATIBILITY
+    SameMajorVersion
+)
+CONFIGURE_PACKAGE_CONFIG_FILE(
+  "${PROJECT_SOURCE_DIR}/cmake/${PROJECT_NAME}Config.cmake.in"
+  "${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
+  INSTALL_DESTINATION
+    ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+INSTALL(
+  EXPORT
+    ${PROJECT_NAME}_Targets
+  FILE
+    ${PROJECT_NAME}Targets.cmake
+  NAMESPACE
+    ${PROJECT_NAME}::
+  DESTINATION
+    ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+INSTALL(
+  DIRECTORY
+    ${PROJECT_SOURCE_DIR}/include/peelo
+  DESTINATION
+    include
+)
+
+IF(PEELO_PROMPT_BUILD_EXAMPLE)
   ADD_SUBDIRECTORY(example)
 ENDIF()
