+++ cmake/PeeloPromptConfig.cmake.in
@@ -1,0 +1,4 @@
+@PACKAGE_INIT@
+
+INCLUDE("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
+CHECK_REQUIRED_COMPONENTS("@PROJECT_NAME@")
